// src/app/shared/generated/model/agentPageResult.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AgentAbstract } from './agentAbstract';


export interface AgentPageResult { 
    totalElements?: number;
    number?: number;
    size?: number;
    totalPages?: number;
    stream?: Array<AgentAbstract>;
}